+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/Implementation/OtlpLogRecordTransformer.cs
@@ -91,18 +91,19 @@ internal sealed class OtlpLogRecordTransformer
             var attributeValueLengthLimit = this.sdkLimitOptions.LogRecordAttributeValueLengthLimit;
             var attributeCountLimit = this.sdkLimitOptions.LogRecordAttributeCountLimit ?? int.MaxValue;
 
-            if (this.experimentalOptions.EmitLogEventAndCategoryAttributes)
+            if (this.experimentalOptions.EmitLogCategoryAttribute)
             {
                 if (!string.IsNullOrEmpty(logRecord.CategoryName))
                 {
                     // TODO:
-                    // 1. Track the following issue, and map CategoryName to Name
-                    // if it makes it to log data model.
-                    // https://github.com/open-telemetry/opentelemetry-specification/issues/2398
-                    // 2. Confirm if this name for attribute is good.
+                    // Decide how to handle CategoryName
+                    // https://github.com/open-telemetry/opentelemetry-dotnet/issues/3491
                     AddStringAttribute(otlpLogRecord, "dotnet.ilogger.category", logRecord.CategoryName, attributeValueLengthLimit, attributeCountLimit);
                 }
+            }
 
+            if (this.experimentalOptions.EmitLogEventAttributes)
+            {
                 if (logRecord.EventId.Id != default)
                 {
                     AddIntAttribute(otlpLogRecord, "event.id", logRecord.EventId.Id, attributeCountLimit);